🚚 Move `lorem` arbitrary into /arbitrary
